김준영 add 페이즈 넘어갈때 카드버리기
