+++ src/classes/models/game.class.js
@@ -206,7 +206,7 @@ class Game {
     return --this.getCharacter(userId).handCardsCount;
   }
 
-  // 카드가 유저의 손, 장비, 디버프, 총 에서 사라질때.
+  // 카드가 유저의 핸드에서 제거될때.
   removeCard(userId, cardType) {
     const handCards = this.getCharacter(userId).handCards;
     const index = handCards.findIndex((card) => card.type === cardType);
@@ -214,6 +214,7 @@ class Game {
     if (index !== -1) {
       handCards[index].count > 1 ? (handCards[index].count -= 1) : handCards.splice(index, 1);
     }
+    this.getCharacter(userId).handCardsCount = handCards.length;
   }
 
   BbangShooterStateInfo(userId, targeId) {
